Move transaction fields to contant file
